+++ PONRF/PonrfClient/PonrfStyles.axaml
@@ -1,12 +1,12 @@
 ﻿<Styles xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <Design.PreviewWith>
-        <Border Padding="20">
-            <!-- Add Controls for Previewer Here -->
-        </Border>
-    </Design.PreviewWith>
+	<Design.PreviewWith>
+		<Border Padding="20">
+			<!-- Add Controls for Previewer Here -->
+		</Border>
+	</Design.PreviewWith>
 
-    <Style Selector="Button.ToolButton">
+	<Style Selector="Button.ToolButton">
 		<Setter Property="Height" Value="45"/>
 		<Setter Property="Width" Value="45"/>
 		<Setter Property="Margin" Value="0,4,0,4"/>
@@ -25,9 +25,9 @@
 	<Style Selector="Button.ButtonWindow">
 		<Setter Property="Margin" Value="4, 10, 4, 10"/>
 	</Style>
-	
+
 	<Style Selector="Grid.MainGrid">
-		<Setter Property="Margin" Value="8"/>
+		<Setter Property="Margin" Value="20"/>
 	</Style>
-	
+
 </Styles>
