+++ PONRF/PonrfClient/Views/ShowAuctionWindow.axaml
@@ -1,0 +1,51 @@
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        x:Class="PonrfClient.Views.ShowAuctionWindow"
+        xmlns:vm="using:PonrfClient.ViewModels"
+		Icon="/Assets/logo3.png"
+		Title="Auction"
+		MaxWidth="500"
+		SizeToContent="Height">
+    <Design.DataContext>
+		<vm:PrivatizedBuildingViewModel/>
+    </Design.DataContext>
+
+	<Window.Styles>
+		<StyleInclude Source="/PonrfStyles.axaml" />
+	</Window.Styles>
+	<Panel>
+		<ExperimentalAcrylicBorder IsHitTestVisible="False">
+			<ExperimentalAcrylicBorder.Material>
+				<ExperimentalAcrylicMaterial
+					BackgroundSource="Digger"
+					TintColor="Black"
+					TintOpacity="1"
+					MaterialOpacity="0.75" />
+			</ExperimentalAcrylicBorder.Material>
+		</ExperimentalAcrylicBorder>
+		<Grid Classes="MainGrid" RowDefinitions="Auto,*">
+			<StackPanel Grid.Row="0" Orientation="Horizontal">
+				<Button Classes="ToolButton" Command="{Binding OnAddAuctionCommand}">
+					<Image Source="/Assets/Add.png" />
+				</Button>
+				<Button Classes="ToolButton" Command="{Binding OnChangeAuctionCommand}">
+					<Image Source="/Assets/Change.png"/>
+				</Button>
+				<Button Classes="ToolButton" Command="{Binding OnDeleteAuctionCommand}">
+					<Image Source="/Assets/Delete.png" />
+				</Button>
+			</StackPanel>
+			<DataGrid Grid.Row="1" AutoGenerateColumns="False"
+					  Items="{Binding	Auctions}"
+					  SelectedItem="{Binding SelectedAuction}">
+				<DataGrid.Columns>
+					<DataGridTextColumn Header="Date" Binding="{Binding Date}"/>
+					<DataGridTextColumn Header="Organizer" Binding="{Binding Organizer}"/>
+				</DataGrid.Columns>
+			</DataGrid>
+		</Grid>
+	</Panel>
+</Window>
