+++ PONRF/PonrfClient/Views/ShowAuctionsWithoutFullSalesWindow.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.ReactiveUI;
+using PonrfClient.ViewModels;
+
+namespace PonrfClient.Views;
+public partial class ShowAuctionsWithoutFullSalesWindow : ReactiveWindow<AuctionsWithoutFullSalesViewModel>
+{
+    public ShowAuctionsWithoutFullSalesWindow()
+    {
+        InitializeComponent();
+    }
+}
