+++ PONRF/PonrfClient/Views/ShowBuildingWindow.axaml
@@ -1,0 +1,56 @@
+<Window xmlns="https://github.com/avaloniaui"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        x:Class="PonrfClient.Views.ShowBuildingWindow"
+        xmlns:vm="using:PonrfClient.ViewModels"
+		Icon="/Assets/logo3.png"
+		Title="Building"
+		MaxWidth="727"
+		SizeToContent="Height">
+    <Design.DataContext>
+		<vm:PrivatizedBuildingViewModel/>
+    </Design.DataContext>
+
+	<Window.Styles>
+		<StyleInclude Source="/PonrfStyles.axaml" />
+	</Window.Styles>
+	<Panel>
+		<ExperimentalAcrylicBorder IsHitTestVisible="False">
+			<ExperimentalAcrylicBorder.Material>
+				<ExperimentalAcrylicMaterial
+					BackgroundSource="Digger"
+					TintColor="Black"
+					TintOpacity="1"
+					MaterialOpacity="0.75" />
+			</ExperimentalAcrylicBorder.Material>
+		</ExperimentalAcrylicBorder>
+		<Grid Classes="MainGrid" RowDefinitions="Auto,*">
+			<StackPanel Grid.Row="0" Orientation="Horizontal">
+				<Button Classes="ToolButton" Command="{Binding OnAddBuildingCommand}">
+					<Image Source="/Assets/Add.png" />
+				</Button>
+				<Button Classes="ToolButton" Command="{Binding OnChangeBuildingCommand}">
+					<Image Source="/Assets/Change.png"/>
+				</Button>
+				<Button Classes="ToolButton" Command="{Binding OnDeleteBuildingCommand}">
+					<Image Source="/Assets/Delete.png" />
+				</Button>
+			</StackPanel>
+			<DataGrid Grid.Row="1" AutoGenerateColumns="False"
+					  Items="{Binding	Buildings}"
+					  SelectedItem="{Binding SelectedBuilding}">
+				<DataGrid.Columns>
+					<DataGridTextColumn Header="Date of build" Binding="{Binding DateOfBuild}"/>
+					<DataGridTextColumn Header="Registration number" Binding="{Binding RegistNum}"/>
+					<DataGridTextColumn Header="District" Binding="{Binding District}"/>
+					<DataGridTextColumn Header="Street" Binding="{Binding Street}"/>
+					<DataGridTextColumn Header="House number" Binding="{Binding HouseNumber}"/>
+					<DataGridTextColumn Header="Area" Binding="{Binding Area}"/>
+					<DataGridTextColumn Header="Floors" Binding="{Binding Floors}"/>
+				</DataGrid.Columns>
+			</DataGrid>
+		</Grid>
+	</Panel>
+</Window>
