+++ PONRF/PonrfClient/Views/ShowTopAuctionWindow.axaml.cs
@@ -1,0 +1,12 @@
+using Avalonia.ReactiveUI;
+using PonrfClient.ViewModels;
+
+namespace PonrfClient.Views;
+public partial class ShowTopAuctionWindow : ReactiveWindow<TopAuctionViewModel>
+{
+    public ShowTopAuctionWindow()
+    {
+        InitializeComponent();
+    }
+
+}
