+++ PONRF/PonrfClient/Views/ShowTopCustomerWindow.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.ReactiveUI;
+using PonrfClient.ViewModels;
+
+namespace PonrfClient.Views;
+public partial class ShowTopCustomerWindow : ReactiveWindow<TopCustomerViewModel>
+{
+    public ShowTopCustomerWindow()
+    {
+        InitializeComponent();
+    }
+}
