+++ PONRF/PonrfClient/Views/ShowViewAllCustomerWindow.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.ReactiveUI;
+using PonrfClient.ViewModels;
+
+namespace PonrfClient.Views;
+public partial class ShowViewAllCustomerWindow : ReactiveWindow<ShowCustomerViewModel>
+{
+    public ShowViewAllCustomerWindow()
+    {
+        InitializeComponent();
+    }
+}
